% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarmas_bus.R
\name{alarmas_bus}
\alias{alarmas_bus}
\title{Genera 2 posibles tipos de alarma: 1) BUS PARADO EN MEDIO DE UNA RUTA + X MINUTOS y 2) AUSENCIA DE DATOS BUS MIENTRAS ESTE ESTABA EFECTUANDO RUTA}
\usage{
alarmas_bus(id_dispositivo, tiempo_maximo_parado)
}
\arguments{
\item{id_dispositivo, tiempo_maximo_parado}{}
}
\value{
json
}
\description{
Genera 2 posibles tipos de alarma: 1) BUS PARADO EN MEDIO DE UNA RUTA + X MINUTOS y 2) AUSENCIA DE DATOS BUS MIENTRAS ESTE ESTABA EFECTUANDO RUTA
}
\examples{
 tiempos_llegada_paradas("b37c2cc0-0350-11ed-b4eb-0d97eeef399c",7)

}
