% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobus_fuera_de_servicio.R
\name{autobus_fuera_de_servicio}
\alias{autobus_fuera_de_servicio}
\title{Asigna línea fuera de servicio cuando el autobús deja de emitir.}
\usage{
autobus_fuera_de_servicio(id_dispositivo)
}
\arguments{
\item{id_dispositivo}{}
}
\value{
json
}
\description{
Asigna línea fuera de servicio cuando el autobús deja de emitir. El trigger es el eventro "active == FALSE" generado en plataforma
}
\examples{
 autobus_fuera_de_servicio("b37c2cc0-0350-11ed-b4eb-0d97eeef399c")

}
