% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo_tiempo_linea.R
\name{calculo_tiempo_linea}
\alias{calculo_tiempo_linea}
\title{Cálculo del tiempo de recorrido de línea y sentido}
\usage{
calculo_tiempo_linea(id_dispositivo, linea, sentido)
}
\arguments{
\item{id_dispositivo, linea, sentido}{}
}
\value{
json
}
\description{
Cálculo del tiempo de recorrido de línea y sentido
}
\examples{
 calculo_tiempo_linea("b37c2cc0-0350-11ed-b4eb-0d97eeef399c",3,1)

}
