% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/festivo_plasencia.R
\name{festivo_plasencia}
\alias{festivo_plasencia}
\title{Comprueba si es festivo y escribe un 1 o 0 en un atributo "festivo" el activo "datos_lineas"}
\usage{
festivo_plasencia()
}
\value{
json
}
\description{
Comprueba si es festivo y escribe un 1 o 0 en un atributo "festivo" el activo "datos_lineas"
}
\examples{
 festivo_plasencia()

}
