% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linea_sin_autobus.R
\name{linea_sin_autobus}
\alias{linea_sin_autobus}
\title{Asigna guiones si no hay ningún autobús circulando en la línea}
\usage{
linea_sin_autobus(linea)
}
\arguments{
\item{linea}{}
}
\value{
json
}
\description{
Asigna guiones si no hay ningún autobús circulando en la línea. Se activa con un cambio de línea
}
\examples{
 linea_sin_autobus(3)

}
