% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencia_paradas_y_tiempos_OD.R
\name{referencia_paradas_y_tiempos_OD}
\alias{referencia_paradas_y_tiempos_OD}
\title{Genera la referencia de paradas en base a activos tipo "parada" de plataforma y la referencia de tiempos de paso de bus (matriz origen destino)}
\usage{
referencia_paradas_y_tiempos_OD(input)
}
\arguments{
\item{input}{}
}
\value{
json
}
\description{
Genera la referencia de paradas en base a activos tipo "parada" de plataforma y la referencia de tiempos de paso de bus (matriz origen destino)
}
\examples{
 referencia_paradas_y_tiempos_OD(1)

}
