% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiempos_llegada_paradas.R
\name{tiempos_llegada_paradas}
\alias{tiempos_llegada_paradas}
\title{Asigna el tiempo de llegada a las marquesinas destino, actualizando el atributos en la plataforma}
\usage{
tiempos_llegada_paradas(id_dispositivo, linea)
}
\arguments{
\item{id_dispositivo, linea}{}
}
\value{
json
}
\description{
Asigna el tiempo de llegada a las marquesinas destino, actualizando el atributos en la plataforma
}
\examples{
 tiempos_llegada_paradas("b37c2cc0-0350-11ed-b4eb-0d97eeef399c",3)

}
